+++ packages/trueforge/tests/unit/apis/mcpOAuth.test.ts
@@ -7,7 +7,8 @@ import { createMcpOAuthRouter } from '../../../src/apis/mcpOAuth';
 import { createMcpServersRouter, createSettingsMcpServersRouter } from '../../../src/apis/mcpServers';
 import { LOCAL_USER_CONTEXT } from '../../../src/auth/identity';
 import configuration from '../../../src/config';
-import { McpServerWithAuthStore, type IMcpServerWithAuthStore } from '../../../src/db/mcpServerStore';
+import { McpServerWithAuthStore } from '../../../src/db/McpServerWithAuthStore';
+import type { IMcpServerWithAuthStore } from '../../../src/db/mcpServerStore';
 import { migrateSqliteToLatest } from '../../../src/db/migrateSqlite';
 import { createSqliteDb } from '../../../src/db/sqlite/client';
 import { SqliteMcpServerStore } from '../../../src/db/sqlite/mcp-server-store/SqliteMcpServerStore';
