+++ packages/trueforge/tests/unit/apis/mcpServers.test.ts
@@ -8,7 +8,8 @@ import { ModelCatalog } from '../../../src/catalog/ModelCatalog';
 import { SandboxCatalog } from '../../../src/catalog/SandboxCatalog';
 import { SkillCatalog } from '../../../src/catalog/SkillCatalog';
 import configuration from '../../../src/config';
-import { McpServerWithAuthStore, type IMcpServerWithAuthStore } from '../../../src/db/mcpServerStore';
+import { McpServerWithAuthStore } from '../../../src/db/McpServerWithAuthStore';
+import type { IMcpServerWithAuthStore } from '../../../src/db/mcpServerStore';
 import { migrateSqliteToLatest } from '../../../src/db/migrateSqlite';
 import { createSqliteDb } from '../../../src/db/sqlite/client';
 import { SqliteMcpServerStore } from '../../../src/db/sqlite/mcp-server-store/SqliteMcpServerStore';
